refactor datatable slightly, make full width
