+++ src/components/Table/@DataTable.scss
@@ -1,0 +1,5 @@
+#data-table{
+}
+#data-table .pagination {
+	float: right;
+}
